Split Marginal model into learning phase
