+++ fonduer/learning/models/__init__.py
@@ -1,0 +1,3 @@
+from fonduer.learning.models.marginal import Marginal
+
+__all__ = ["Marginal"]
